+++ src/routes/Part/GetPart.ts
@@ -25,12 +25,12 @@ const partDao = new PartDao();
       const part = new Part(
           partData.name.S,
           partData.PartType.S,
-          parseInt(partData.Speed.N),
-          parseInt(partData.Acceleration.N),
-          parseInt(partData.Weight.N),
-          parseInt(partData.Handling.N),
-          parseInt(partData.Traction.N),
-          parseInt(partData.MiniTraction.N)
+          parseFloat(partData.Speed.N),
+          parseFloat(partData.Acceleration.N),
+          parseFloat(partData.Weight.N),
+          parseFloat(partData.Handling.N),
+          parseFloat(partData.Traction.N),
+          parseFloat(partData.MiniTraction.N)
       );
       return res.status(OK).json(part);
   }
